+++ Steps/Azure/ServiceBus/test/ReadMessageFromStepOptionsTests.cs
@@ -1,7 +1,7 @@
 ﻿// <copyright file="ReadMessageFromAzureStepTests.cs" company="Chris Trout">
 // MIT License
 //
-// Copyright(c) 2019-2020 Chris Trout
+// Copyright(c) 2019-2022 Chris Trout
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -36,33 +36,34 @@ namespace MyTrout.Pipelines.Steps.Azure.ServiceBus.Tests
         public void Constructs_ReadMessageFromStepOptions_Successfully()
         {
             // arrange
-            List<string> applicationProperties = new List<string>() { "UserId", "MessageId" };
             string azureServiceBusConnectionString = "ConnectionString";
             int batchSize = 100;
             int deliveryAttemptsBeforeDeadLetter = 2;
             string entityPath = "topic/subscription";
-            TimeSpan timeToWaitBetweenMessageChecks = new TimeSpan(0, 0, 1);
-            TimeSpan timeToWaitForNewMessage = new TimeSpan(0, 1, 0);
+            string messagePrefix = "CHANGED_";
+            var timeToWaitBetweenMessageChecks = new TimeSpan(0, 0, 1);
+            var timeToWaitForNewMessage = new TimeSpan(0, 1, 0);
+
 
             // act
             var result = new ReadMessageFromAzureOptions
             {
-                ApplicationProperties = applicationProperties,
                 AzureServiceBusConnectionString = azureServiceBusConnectionString,
                 BatchSize = batchSize,
                 DeliveryAttemptsBeforeDeadLetter = deliveryAttemptsBeforeDeadLetter,
                 EntityPath = entityPath,
+                MessageContextItemsPrefix = messagePrefix,
                 TimeToWaitBetweenMessageChecks = timeToWaitBetweenMessageChecks,
                 TimeToWaitForNewMessage = timeToWaitForNewMessage
             };
 
             // assert
             Assert.IsNotNull(result);
-            Assert.AreEqual(applicationProperties, result.ApplicationProperties);
             Assert.AreEqual(azureServiceBusConnectionString, result.AzureServiceBusConnectionString);
             Assert.AreEqual(batchSize, result.BatchSize);
             Assert.AreEqual(deliveryAttemptsBeforeDeadLetter, result.DeliveryAttemptsBeforeDeadLetter);
             Assert.AreEqual(entityPath, result.EntityPath);
+            Assert.AreEqual(messagePrefix, result.MessageContextItemsPrefix);
             Assert.AreEqual(timeToWaitBetweenMessageChecks, result.TimeToWaitBetweenMessageChecks);
             Assert.AreEqual(timeToWaitForNewMessage, result.TimeToWaitForNewMessage);